Fusion instrument selection is in (not during performance)
The widget can be used to change the fusion instrument, this is a little glitchy and will only apply the next time the performer is initialized, which is not streamlined at all right now. The idea is for the fusion patch to be updated during playback if the session is playing (also during gameplay).