Change default output to go through bus when playing from musicactor
This ensures that spatial outputs and the main output are not out of phase while maintaining almost-perfect video sync.

In theory the audio should only be offset by one buffer, which should be less than a video frame, probably, maybe changing the call back buffer size can improve this, I need to test, in theory it's also possible to simply delay the video.