propogate metasound editor custom icons to m2sound editor
that was a fun exploration, currently only applies to the 'patch' nodes because I guess I should refactor some stuff so that it applies to everything (in fact we only need one single type of UEdGraphNode - follow the way things are done in the metasound editor). So really this can only be seen in the midi stream inputs/outputs of our patch nodes which now display the harmonix icon.